fix(frontend): mysql db重命名 #10279 # Reviewed, transaction id: 42905
