+++ dbm-ui/frontend/src/common/const/ticketTypes.ts
@@ -20,7 +20,6 @@ export enum TicketTypes {
   MYSQL_HA_DISABLE = 'MYSQL_HA_DISABLE',
   MYSQL_HA_ENABLE = 'MYSQL_HA_ENABLE',
   MYSQL_HA_FULL_BACKUP = 'MYSQL_HA_FULL_BACKUP',
-  MYSQL_HA_RENAME_DATABASE = 'MYSQL_HA_RENAME_DATABASE',
   MYSQL_HA_TRUNCATE_DATA = 'MYSQL_HA_TRUNCATE_DATA',
   MYSQL_IMPORT_SQLFILE = 'MYSQL_IMPORT_SQLFILE',
   MYSQL_INSTANCE_CLONE_RULES = 'MYSQL_INSTANCE_CLONE_RULES',
@@ -34,6 +33,7 @@ export enum TicketTypes {
   MYSQL_PROXY_ADD = 'MYSQL_PROXY_ADD',
   MYSQL_PROXY_SWITCH = 'MYSQL_PROXY_SWITCH',
   MYSQL_PROXY_UPGRADE = 'MYSQL_PROXY_UPGRADE', // MySQL Proxy 升级
+  MYSQL_RENAME_DATABASE = 'MYSQL_RENAME_DATABASE',
   MYSQL_RESTORE_LOCAL_SLAVE = 'MYSQL_RESTORE_LOCAL_SLAVE',
   MYSQL_RESTORE_SLAVE = 'MYSQL_RESTORE_SLAVE',
   MYSQL_ROLLBACK_CLUSTER = 'MYSQL_ROLLBACK_CLUSTER',
@@ -41,7 +41,6 @@ export enum TicketTypes {
   MYSQL_SINGLE_DESTROY = 'MYSQL_SINGLE_DESTROY',
   MYSQL_SINGLE_DISABLE = 'MYSQL_SINGLE_DISABLE',
   MYSQL_SINGLE_ENABLE = 'MYSQL_SINGLE_ENABLE',
-  MYSQL_SINGLE_RENAME_DATABASE = 'MYSQL_SINGLE_RENAME_DATABASE',
   MYSQL_SINGLE_TRUNCATE_DATA = 'MYSQL_SINGLE_TRUNCATE_DATA',
   MYSQL_SLAVE_MIGRATE_UPGRADE = 'MYSQL_SLAVE_MIGRATE_UPGRADE', // MySQL 从库迁移升级
 }
