+++ dbm-ui/frontend/src/services/model/ticket/details/mysql/index.ts
@@ -14,7 +14,6 @@ export * from './haDestroy';
 export * from './haDisable';
 export * from './haEnable';
 export * from './haFullBackup';
-export * from './haRenameDatabase';
 export * from './importSqlFile';
 export * from './instanceCloneRules';
 export * from './localUpgrade';
@@ -27,6 +26,7 @@ export * from './partition';
 export * from './proxyAdd';
 export * from './proxySwitch';
 export * from './proxyUpgrade';
+export * from './renameDataBase';
 export * as ResourcePool from './resource-pool/index';
 export * from './restoreLocalSlave';
 export * from './restoreSlave';
@@ -35,5 +35,4 @@ export * from './singleApply';
 export * from './singleDestroy';
 export * from './singleDisable';
 export * from './singleEnable';
-export * from './singleRenameDatabase';
 export * from './truncateData';
