+++ dbm-ui/frontend/src/services/model/ticket/details/mysql/renameDataBase.ts
@@ -1,11 +1,14 @@
 import type { DetailBase, DetailClusters } from '../common';
 
-export interface SingleRenameDatabase extends DetailBase {
+/**
+ *  Mysql DB 重命名
+ */
+
+export interface RenameDataBase extends DetailBase {
   clusters: DetailClusters;
   force: boolean;
   infos: {
     cluster_id: number;
-    force: boolean;
     from_database: string;
     to_database: string;
   }[];
